update migration setup

+++ src/config/database/migrations/1782000000000-PaymentSelectedBankAccount.ts
@@ -6,6 +6,10 @@ export class PaymentSelectedBankAccount1782000000000
   name = 'PaymentSelectedBankAccount1782000000000';
 
   public async up(queryRunner: QueryRunner): Promise<void> {
+    if (!(await queryRunner.hasTable('payments'))) {
+      return;
+    }
+
     await queryRunner.query(`
       ALTER TABLE "payments"
       ADD COLUMN IF NOT EXISTS "selectedBankAccount" jsonb NULL
@@ -13,6 +17,10 @@ export class PaymentSelectedBankAccount1782000000000
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
+    if (!(await queryRunner.hasTable('payments'))) {
+      return;
+    }
+
     await queryRunner.query(`
       ALTER TABLE "payments"
       DROP COLUMN IF EXISTS "selectedBankAccount"
